backend/requirements.txt
fastapi==0.115.2
uvicorn[standard]==0.30.6
SQLAlchemy==2.0.36
PyMySQL==1.1.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==3.2.2
python-multipart==0.0.12
pydantic==2.9.2
python-dotenv==1.0.1
email-validator==2.2.0
Pillow==10.4.0
